Add active_window_ignore_refnum option (bug #667)
With active_window_ignore_refnum = ON, the current behavior for the
active_window key (meta-a by default) is preserved: windows are cycled in the
order of most recent activity, highest activity first.

With active_window_ignore_refnum = OFF, the old behavior is used: windows are
cycled in the order of most recent activity, where ties of equally high activity
are broken by refnums. Windows with lower refnums and equal activity will be
chosen first.